% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_propagation.R
\name{creep_quartz_analytic}
\alias{creep_quartz_analytic}
\title{Algebraic error propagation of flow laws}
\usage{
creep_quartz_analytic(
  stress,
  temperature,
  fugacity = NULL,
  grainsize = NULL,
  pressure = NULL,
  model = c("Hirth2001", "Paterson1990", "Kronenberg1984", "Luan1992", "Gleason1995",
    "Gleason1995_melt", "Rutter2004", "Fukuda2018_LT", "Fukuda2018_HT", "Richter2018",
    "Lu2019", "Tokle2019_LT", "Tokle2019_HT", "Lusk2021_LP", "Lusk2021_HP")
)
}
\arguments{
\item{stress}{Differential stress in MPa or \code{units} object}

\item{temperature}{Temperature in Kelvin or \code{units} object}

\item{fugacity}{Water fugacity in MPa or \code{units} object}

\item{grainsize}{Grain size in \eqn{\mu}m or \code{units} object}

\item{pressure}{Pressure in MPa or \code{units} object}

\item{model}{character specifying the flow law to be used:
\describe{
\item{\code{"Hirth2001"}}{Hirth and Tullis (2001), dislocation creep}
\item{\code{"Paterson1990"}}{Paterson and Luan (1990): dislocation creep; axial compression}
\item{\code{"Kronenberg1984"}}{Kronenberg and Tullis (1984): deformation mechanism: dislocation creep and grain-size sensitive creep; strain geometry: axial compression}
\item{\code{"Luan1992"}}{Luan and Paterson (1990): dislocation creep; axial compression}
\item{\code{"Gleason1995"}}{Gleason and Tullis (1995): dislocation creep; axial compression}
\item{\code{"Gleason1995_melt"}}{Gleason and Tullis (1995): dislocation creep; axial compression; 1–2\% melt}
\item{\code{"Rutter2004"}}{Rutter and Brodie (2004b): dislocation creep; axial compression}
\item{\code{"Fukuda2018_LT"}}{Fukada et al. (2018): dislocation creep; axial compression; low temperatures (600–750 °C)}
\item{\code{"Fukuda2018_HT"}}{Fukada et al. (2018): dislocation creep and grain-size sensitive creep; axial compression; high temperatures (800–950 °C)}
\item{\code{"Richter2018"}}{Richter et al. (2018): dislocation creep and grain-size sensitive creep; general shear; 800–1000 °C}
\item{\code{"Lu2019"}}{Lu and Jiang (2019): dislocation creep}
\item{\code{"Tokle2019_HT"}}{Tokle et al. (2019): dislocation creep and grain-size sensitive creep; high temperatures/low stress}
\item{\code{"Tokle2019_LT"}}{Tokle et al. (2019): dislocation creep and grain-size sensitive creep; low temperature/high stress}
\item{\code{"Lusk2021_LP"}}{Lusk et al. (2021): dislocation-dominated creep in wet quartz, for low pressures (≤560 MPa)}
\item{\code{"Lusk2021_HP"}}{Lusk et al. (2021): dislocation-dominated creep in wet quartz, for high pressures (700–1600 MPa)}
}}
}
\value{
list.\describe{
\item{\code{"e_best"}}{Strain rate (1/s)}
\item{\code{"sd_e_range"}}{Range of 1s standard deviation  (1/s)}
\item{\code{"var_log_e_total"}}{Variance in natural log-space}
\item{\verb{"sd_log_e_total}}{Standard deviation in natural log-space}
\item{\code{"var_log_e"}}{Individual variance components in natural log-space}
}
}
\description{
Uses 1st order Taylor expansion (linearization) of the natural log-space flow law
}
\examples{
stress <- units::set_units(100, MPa)
temperature <- units::set_units(300, degC)
pressure <- units::set_units(400, MPa)
fugacity <- ps_fugacity(pressure, temperature)

creep_quartz_analytic(
  stress = stress, temperature = temperature, fugacity = fugacity, pressure = pressure,
  model = "Hirth2001"
)
}
\seealso{
\code{\link[=creep_quartz]{creep_quartz()}}
}
