% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peierl.R
\name{peierls_creep}
\alias{peierls_creep}
\title{Peierls Creep}
\usage{
peierls_creep(
  stress,
  temperature,
  model = c("Goetze1979", "Demouchy2013", "Mei2010"),
  sim = 1e+06
)
}
\arguments{
\item{stress}{Differential stress in MPa or \code{units} object}

\item{temperature}{Temperature in Kelvin or \code{units} object}

\item{model}{character. Model to be used, one of \code{"Goetze1979"}
(Goetze and Evans, 1979), \code{"Demouchy2013"} (Demouchy et al, 2013), and
\code{"Mei2010"} (Mei et al., 2010)}

\item{sim}{integer. Number of Monte Carlo simulations.}
}
\value{
Strain rate in 1/s
}
\description{
Strain rate for low-temperature, plastic deformation mechanism that corresponds to
dislocation glide at high stress (differential stress larger than 200 MPa).
}
\examples{
peierls_creep(300, 500, model = "Goetze1979")
peierls_creep(300, 500, model = "Demouchy2013")
peierls_creep(300, 500, model = "Mei2010") |> summary()
}
