% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strainrate.R
\name{strain_rate}
\alias{strain_rate}
\title{Strain Rate}
\usage{
strain_rate(
  stress,
  temperature,
  fugacity = NULL,
  grainsize = NULL,
  pressure = NULL,
  sim = 1e+06,
  model = c("Hirth2001", "Paterson1990", "Kronenberg1984", "Luan1992", "Gleason2004",
    "Rutter2004", "Fukuda2018", "Richter2018", "Lu2019", "Tokle2019", "Lusk2021")
)
}
\arguments{
\item{stress}{Deviatoric stress in MPa or \code{units} object}

\item{temperature}{Temperature in Kelvin or \code{units} object}

\item{fugacity}{Water fugacity in MPa or \code{units} object}

\item{grainsize}{Grain size in \eqn{\mu}m or \code{units} object}

\item{pressure}{Pressure in MPa or \code{units} object}

\item{sim}{non-negative number. Number of Monte Carlo simulations}

\item{model}{character specifying the flow law to be used: \describe{
\item{\code{"Hirth2001"}}{Hirth and Tullis (2001), dislocation creep}
\item{\code{"Paterson1990"}}{Paterson and Luan (1990): dislocation creep; axial compression}
\item{\code{"Kronenberg1984"}}{Kronenberg and Tullis (1984): deformation mechanism: dislocation creep and grain-size sensitive creep; strain geometry: axial compression}
\item{\code{"Luan1992"}}{Luan and Paterson (1990): dislocation creep; axial compression}
\item{\code{"Gleason2004"}}{Gleason and Tullis (1995): dislocation creep; axial compression}
\item{\code{"Rutter2004"}}{Rutter and Brodie (2004): dislocation creep; axial compression}
\item{\code{"Fukuda2018"}}{Fukada et al. (2018): dislocation creep; axial compression}
\item{\code{"Fukuda2018_2"}}{Fukada et al. (2018): dislocation creep and grain-size sensitive creep; axial compression}
\item{\code{"Richter2018"}}{Richter et al. (2018): dislocation creep and grain-size sensitive creep; general shear}
\item{\code{"Lu2019"}}{Lu and Jiang (2019): dislocation creep}
\item{\code{"Tokle2019"}}{Tokle et al. (2019): dislocation creep and grain-size sensitive creep}
\item{\code{"Tokle2019_2"}}{Tokle et al. (2019): dislocation creep and grain-size sensitive creep}
\item{\code{"Lusk2021"}}{Lusk et al. (2021): dislocation-dominated creep in wet quartz, for low pressures (less than 560 MPa)}
\item{\code{"Lusk2021_HP"}}{Lusk et al. (2021): dislocation-dominated creep in wet quartz, for high pressures (700-1600 MPa)}
}}
}
\value{
list. Strain rate in 1/s. If Monte Carlo Simulation was used, see \code{\link[=mc_stats]{mc_stats()}} for detailed description of output.
The flow laws produce log-normal distributed estimates considering the uncertainties in the parameter. Hence it is recommended to report the median (or geometric mean), and the interpercentile range.
}
\description{
Calculates strain rate from stress, temperature, and grain size.
Uses Monte Carlo sampling for propagating parameter uncertainties in flow model.
}
\details{
General flow law: \deqn{\dot{\epsilon} = A \sigma^n d^m f_{H_2O}^r \, e^{\left({\frac{-H}{RT}}\right)}}
}
\examples{
set.seed(20250411)
stress <- units::set_units(100, MPa)
temperature <- units::set_units(300, degC)
pressure <- units::set_units(400, MPa)
fugacity <- ps_fugacity(pressure, temperature)
strain_rate(stress = stress, temperature = temperature, model = "Paterson1990")
strain_rate(stress = stress, temperature = temperature, fugacity = fugacity, model = "Hirth2001")
strain_rate(stress = stress, temperature = temperature, fugacity = fugacity, model = "Rutter2004")
strain_rate(stress = stress, temperature = temperature, fugacity = fugacity, model = "Lu2019")
}
\references{
Hirth, G., Teyssier, C., & Dunlap, W. J. (2001). An evaluation of quartzite flow laws based on comparisons between experimentally and naturally deformed rocks. International Journal of Earth Sciences, 90(1), 77–87. https://doi.org/10.1007/s005310000152

Tokle, L., Hirth, G., & Behr, W. M. (2019). Flow laws and fabric transitions in wet quartzite. Earth and Planetary Science Letters, 505, 152–161. https://doi.org/10.1016/j.epsl.2018.10.017

Lusk, A. D. J., Platt, J. P., & Platt, J. A. (2021). Natural and Experimental Constraints on a Flow Law for Dislocation‐Dominated Creep in Wet Quartz. Journal of Geophysical Research: Solid Earth, 126(5), 1–25. https://doi.org/10.1029/2020JB021302
}
\seealso{
\code{\link[units:units]{units::set_units()}}
}
