% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grainsize.R
\name{subgrainsize_piezometry}
\alias{subgrainsize_piezometry}
\title{Subgrain‐Size Piezometer Calibrated for EBSD}
\usage{
subgrainsize_piezometry(
  lambda,
  sd = NULL,
  calibrated = TRUE,
  min = c("q", "fo90", "fo50"),
  sim = 1e+06
)
}
\arguments{
\item{lambda}{mean line intercept length in \eqn{\mu}m or \code{units} object.}

\item{sd}{(optional) Standard deviation of \code{lambda}}

\item{calibrated}{logical. Whether the calibration of Holyoke and Kronenberg (2010) is considered or not.}

\item{min}{character. The Mineral uses. one of \code{"q"} for quartz, \code{"fo90"} for Olive with 90\% Forsterite, or \code{"fo50"} for Olivine with 50\% Forsterite.}

\item{sim}{non-negative integer. Number of Monte Carlo simulations}
}
\value{
list. Differential stress in MPa. If Monte Carlo Simulation was used,
and object of class \code{"MC_sim"} is returned (see \code{\link[=mc_stats]{mc_stats()}} for detailed description of output).
The piezometer produce log-normal distributed estimates considering the
uncertainties in the equation parameter. Hence it is recommended to report
the median (or geometric mean), and the interpercentile range.
}
\description{
Calculates differential stress using the Subgrain‐Size Piezometer of Goddard et al. (2020).
Uses Monte Carlo sampling for propagating parameter uncertainties in flow model.
}
\details{
The sub-grain size piezometer is
\deqn{\frac{\lambda}{b} = 10^a \left(\frac{\sigma}{\mu}\right)^b}
where \eqn{\lambda} is the mean line intercept length, \eqn{b} is the Burgers vector,
\eqn{\sigma} is the differential stress, \eqn{\mu} is the shear modulus, and
\eqn{a} and \eqn{b} are the empirical exponents.
}
\examples{
set.seed(20250411)
subgrainsize_piezometry(9, min = "fo50") # 420 MPa
subgrainsize_piezometry(18, min = "q") # 240 MPa
}
\references{
Goddard, R. M., Hansen, L. N., Wallis, D., Stipp, M., Holyoke,
C. W., Kumamoto, K. M., & Kohlstedt, D. L. (2020). A Subgrain‐Size Piezometer
Calibrated for EBSD. Geophysical Research Letters, 47(23).
https://doi.org/10.1029/2020GL090056
}
\seealso{
\code{\link[=grainsize_piezometry]{grainsize_piezometry()}}, \code{\link[units:units]{units::set_units()}}
}
