% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{summary-MCS}
\alias{summary-MCS}
\alias{summary.MCS_log}
\alias{summary.MCS}
\title{Estimates from Monte Carlo Simulation}
\usage{
\method{summary}{MCS_log}(object, unit = NULL, ...)

\method{summary}{MCS}(object, unit = NULL, ...)
}
\arguments{
\item{object}{numeric vector of class \code{"MCS"} or \code{"MCS_log"}. The values
from n Monte Carlo Simulations}

\item{unit}{(optional) object of class \code{units} or \code{symbolic_units}, or in
the case of \code{set_units} expression with symbols.}

\item{...}{additional arguments affecting the summary produced.}
}
\value{
a list.
If class of \code{object} is \code{"MCS"}, the list contains the following elements:
\describe{
\item{\code{median}}{median of the Monte Carlo simulations}
\item{\code{ir.95}}{the 95\% and 68\% interpercentile range}
\item{\code{ir.68}}{the 68\% interpercentile range}
\item{\code{mean}}{arithmetic mean the Monte Carlo simulations}
\item{\code{sd}}{1\eqn{\sigma} standard deviation about the mean}
\item{\code{conf.int}}{95\% confidence intverval about the mean}
\item{\code{"var"}}{Variance}
\item{\code{stderr}}{standard error}
\item{\code{t.test}}{Statistic and p-value of the Student's t-test}
\item{\code{n}}{Number of samples}
}

If class of \code{object} is \code{"MCS_log"}, the list contains the following elements:
\describe{
\item{\code{median}}{median of the Monte Carlo simulations}
\item{\code{ir.95}}{the 95\% and 68\% interpercentile range}
\item{\code{ir.68}}{the 68\% interpercentile range}
\item{\code{mean}}{geometric mean the Monte Carlo simulations}
\item{\code{sd}}{1\eqn{\sigma} range about the mean}
\item{\code{sd2}}{2\eqn{\sigma} range about the mean}
\item{\code{conf.int}}{95\% confidence intverval about the mean}
\item{\code{var.log}}{Log-variance}
\item{\code{stderr.log}}{standard error of \code{log(samples)}}
\item{\code{t.test}}{Statistic and p-value of the Student's t-test of \code{log(samples)}}
\item{\code{n}}{Number of samples}
}
Values will be in the unit specified by parameter \code{unit} or be equal to the
unit of \code{x} if \code{x} is a \code{units} object.
}
\description{
Estimates from Monte Carlo Simulation
}
\details{
Equations of the form \eqn{X = A b^{n \pm \sigma}} create
non-normal, left-skewed distributions (e.g. flow laws, and grain-size
piezometers).
Thus, it is recommended to report median and percentiles instead of mean,
standard deviation and confidence intervals.
}
\examples{
set.seed(20250411)
MC_res <- grainsize_piezometry(12.2)
summary(MC_res)

n <- 100
temperature <- units::set_units(rnorm(n, 300, 25), degC)
pressure <- units::set_units(rnorm(n, 400, 50), MPa)
MC_res2 <- ps_fugacity(pressure, temperature) # 37 MPa
summary(MC_res2)
}
