% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{summary.MCS_log}
\alias{summary.MCS_log}
\title{Estimates from Monte Carlo Simulation}
\usage{
\method{summary}{MCS_log}(object, unit = NULL, ...)
}
\arguments{
\item{object}{numeric vector of class \code{"MCS_log"}. The values from n Monte Carlo Simulations}

\item{unit}{(optional) object of class \code{units} or \code{symbolic_units}, or in the case of \code{set_units} expression with symbols.}

\item{...}{additional arguments affecting the summary produced.}
}
\value{
a list with
\describe{
\item{\code{median}}{median of the Monte Carlo simulations}
\item{\code{mean}}{geometric mean the Monte Carlo simulations}
\item{\code{stderr_log}}{standard error of \code{log(samples)}}
\item{\code{ir_95}}{the 95\% and 68\% interpercentile range}
\item{\code{ir_68}}{the 68\% interpercentile range}
\item{\code{samples}}{the Monte Carlo simulation}
}
Values will be in the unit specified by parameter \code{unit} or be equal to the unit of \code{x} if \code{x} is a \code{units} object.
}
\description{
Estimates from Monte Carlo Simulation
}
\details{
Equations of the form \eqn{X = A b^{n \pm \sigma}} create non-normal, left-skewed distributions.
Thus, it is recommended to report median and percentiles instead of mean, standard deviation and confidence intervals.
}
\examples{
set.seed(20250411)
MC_res <- grainsize_piezometry(12.2)
summary(MC_res)
}
