% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strainrate.R
\name{creep_quartz}
\alias{creep_quartz}
\title{Strain Rates for Creep in Quartz}
\usage{
creep_quartz(
  stress,
  temperature,
  fugacity = NULL,
  grainsize = NULL,
  pressure = NULL,
  sim = 1e+06,
  model = c("Hirth2001", "Paterson1990", "Kronenberg1984", "Luan1992", "Gleason1995",
    "Gleason1995_melt", "Rutter2004", "Fukuda2018_LT", "Fukuda2018_HT", "Richter2018",
    "Lu2019", "Tokle2019_LT", "Tokle2019_HT", "Lusk2021", "Lusk2021_HP")
)
}
\arguments{
\item{stress}{Differential stress in MPa or \code{units} object}

\item{temperature}{Temperature in Kelvin or \code{units} object}

\item{fugacity}{Water fugacity in MPa or \code{units} object}

\item{grainsize}{Grain size in \eqn{\mu}m or \code{units} object}

\item{pressure}{Pressure in MPa or \code{units} object}

\item{sim}{non-negative number. Number of Monte Carlo simulations}

\item{model}{character specifying the flow law to be used:
\describe{
\item{\code{"Hirth2001"}}{Hirth and Tullis (2001), dislocation creep}
\item{\code{"Paterson1990"}}{Paterson and Luan (1990): dislocation creep; axial compression}
\item{\code{"Kronenberg1984"}}{Kronenberg and Tullis (1984): deformation mechanism: dislocation creep and grain-size sensitive creep; strain geometry: axial compression}
\item{\code{"Luan1992"}}{Luan and Paterson (1990): dislocation creep; axial compression}
\item{\code{"Gleason1995"}}{Gleason and Tullis (1995): dislocation creep; axial compression}
\item{\code{"Gleason1995_melt"}}{Gleason and Tullis (1995): dislocation creep; axial compression; 1-2 percent melt}
\item{\code{"Rutter2004"}}{Rutter and Brodie (2004): dislocation creep; axial compression}
\item{\code{"Fukuda2018_LT"}}{Fukada et al. (2018): dislocation creep; axial compression}
\item{\code{"Fukuda2018_HT"}}{Fukada et al. (2018): dislocation creep and grain-size sensitive creep; axial compression; HT-fit}
\item{\code{"Richter2018"}}{Richter et al. (2018): dislocation creep and grain-size sensitive creep; general shear; between 800 and 1000 degree Celsius}
\item{\code{"Lu2019"}}{Lu and Jiang (2019): dislocation creep}
\item{\code{"Tokle2019_HT"}}{Tokle et al. (2019): dislocation creep and grain-size sensitive creep; high temperatures/low stress}
\item{\code{"Tokle2019_LT"}}{Tokle et al. (2019): dislocation creep and grain-size sensitive creep; low temperature/high stress}
\item{\code{"Lusk2021"}}{Lusk et al. (2021): dislocation-dominated creep in wet quartz, for low pressures (less than 560 MPa)}
\item{\code{"Lusk2021_HP"}}{Lusk et al. (2021): dislocation-dominated creep in wet quartz, for high pressures (700-1600 MPa)}
}}
}
\value{
list. Strain rate in 1/s. If Monte Carlo Simulation was used, and
object of class \code{"MCS"} is returned (see \code{\link[=summary]{summary()}} for detailed description of output).
The flow laws produce log-normal distributed estimates considering the
uncertainties in the parameter. Hence it is recommended to report the median
(or geometric mean), and the interpercentile range.
}
\description{
Calculates strain rates of deforming quartz from stress, temperature, and grain size from
experimentally determined creep law parameters.
Monte Carlo sampling is used for propagating parameter uncertainties in to creep estimate.
}
\details{
General flow law giving the strain rate is  \deqn{\dot{\epsilon} = A \sigma^n d^m f_{H_2O}^r \, e^{\left({\frac{-H}{RT}}\right)}}

where \eqn{\sigma} is the differential stress, \eqn{d} is the grain size,
\eqn{f_{H_2O}} is the water fugacity, \eqn{T} is the temperature,
\eqn{H} is the enthalpy, and \eqn{R} is the ideal gas constant. The flow
parameters are the prefactor \eqn{A}, and the exponents \eqn{n}, \eqn{m}, and \eqn{r}.

To propagate the uncertainties of the flow parameters Monte Carlo simulation is used here.
\itemize{
\item{If the flow law parameters are given by a mean value and a marginal error (\eqn{\mu \pm z}),
the Monte Carlo simulation assumes a normal distribution given by
\eqn{X = N\left(\mu, \sigma\right)}, where \eqn{\mu} is the mean and \eqn{\sigma}
is the standard deviation of the mean (\eqn{\sigma = \text{z}/1.96}).}
\item{If the parameter is given by a range of possible values \eqn{\left[x_\text{min}, x_\text{max}\right]},
the Monte Carlo simulation assumes an uniform distribution given by \eqn{X = U\left(x_\text{min}, x_\text{max}\right)}.}
}
}
\examples{
set.seed(20250411)
stress <- units::set_units(100, MPa)
temperature <- units::set_units(300, degC)
pressure <- units::set_units(400, MPa)
fugacity <- ps_fugacity(pressure, temperature)

creep_quartz(stress = stress, temperature = temperature, model = "Paterson1990")
creep_quartz(stress = stress, temperature = temperature, fugacity = fugacity, model = "Hirth2001")
creep_quartz(stress = stress, temperature = temperature, fugacity = fugacity, model = "Rutter2004")
creep_quartz(stress = stress, temperature = temperature, fugacity = fugacity, model = "Lu2019")
}
\references{
Fukuda, J., Holyoke, C. W., & Kronenberg, A. K. (2018). Deformation of Fine‐Grained Quartz Aggregates by Mixed Diffusion and Dislocation Creep. Journal of Geophysical Research: Solid Earth, 123(6), 4676-4696. \doi{10.1029/2017JB015133}

Gleason, G. C., & Tullis, J. (1995). A flow law for dislocation creep of quartz aggregates determined with the molten salt cell. Tectonophysics, 247(1-4), 1-23. \doi{10.1016/0040-1951(95)00011-B}

Hirth, G., Teyssier, C., \& Dunlap, W. J. (2001). An evaluation of quartzite flow laws based on comparisons between experimentally and naturally deformed rocks. International Journal of Earth Sciences, 90(1), 77-87. \doi{10.1007/s005310000152}

Tokle, L., Hirth, G., \& Behr, W. M. (2019). Flow laws and fabric transitions in wet quartzite. Earth and Planetary Science Letters, 505, 152-161. \doi{10.1016/j.epsl.2018.10.017}

Lu, L. X., \& Jiang, D. (2019). Quartz Flow Law Revisited: The Significance of Pressure Dependence of the Activation Enthalpy. Journal of Geophysical Research: Solid Earth, 124(1), 241–256. \doi{10.1029/2018JB016226}

Lusk, A. D. J., Platt, J. P., \& Platt, J. A. (2021). Natural and Experimental Constraints on a Flow Law for Dislocation‐Dominated Creep in Wet Quartz. Journal of Geophysical Research: Solid Earth, 126(5), 1-25. \doi{10.1029/2020JB021302}

Richter, B., Stünitz, H., \& Heilbronner, R. (2018). The brittle-to-viscous transition in polycrystalline quartz: An experimental study. Journal of Structural Geology, 114(September 2017), 1-21. \doi{10.1016/j.jsg.2018.06.005}
}
\seealso{
\code{\link[units:units]{units::set_units()}} to set up \code{units} objects; \code{\link[=summary.MCS_log]{summary.MCS_log()}} for statistical parameters of Monte Carlo samples
}
