% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grainsize.R
\name{grainsize_piezometry}
\alias{grainsize_piezometry}
\title{Grain size piezometry}
\usage{
grainsize_piezometry(
  d,
  sd = NULL,
  method = c("Stipp-reg2-3", "Stripp-reg1", "Cross-1", "Cross-sliding"),
  sim = 1e+06
)
}
\arguments{
\item{d}{numeric. Grain size in micrometer or units object}

\item{sd}{(optional) numeric. Standard deviation of \code{d}}

\item{method}{character. One of
\describe{
\item{Stipp-reg2-3"}{Piezometer for dislocation creep regime 2 and 3 after Stipp and Tullis (2003)}
\item{"Stripp-reg1"}{Piezometer for dislocation creep regime 1 after Stipp and Tullis (2003)}
\item{"Cross-1"}{Piezmeter after Cross et al. (2017) for 1 um step size resolution in EBSD data}
\item{"Cross-sliding"}{Sliding resolution piezometer after Cross et al. 2017. According to authors, more accurately estimates stress in fine-grained (<10 μm) samples}
}}

\item{sim}{non-negative integer. Number of Monte Carlo simulations}
}
\value{
list. Stress in MPa given as median of the Monte Carlo simulations \code{median}, mean (\code{mean}), standard error range (\code{sde}), 95\% and 68\% interpercentile range (\code{ir_95} and \code{ir_68}), and the Monte Carlo simulation (\code{samples}).
}
\description{
Grain size piezometry
}
\examples{
set.seed(20250411)
grainsize_piezometry(12.2) # 92
grainsize_piezometry(31) # 44
}
\references{
Stipp, M., & Tullis, J. (2003). The recrystallized grain size piezometer for
quartz. Geophysical Research Letters, 30(21), 1–5. https://doi.org/10.1029/2003GL018444

Cross, A. J., Prior, D. J., Stipp, M., & Kidder, S. (2017). The
recrystallized grain size piezometer for quartz: An EBSD-based calibration.
Geophysical Research Letters, 44(13), 6667–6674. https://doi.org/10.1002/2017GL073836
}
