% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{mc_stats}
\alias{mc_stats}
\title{Estimates from Monte Carlo Simulation}
\usage{
mc_stats(x, unit = NULL)
}
\arguments{
\item{x}{numeric vector of vector of class \code{units}. The values from n Monte Carlo Simulations}

\item{unit}{(optional) object of class \code{units} or \code{symbolic_units}, or in the case of \code{set_units} expression with symbols.}
}
\value{
\code{'MC_sim'} object, i.e. a list.
\describe{
\item{\code{median}}{median of the Monte Carlo simulations}
\item{\code{mean}}{geometric mean the Monte Carlo simulations}
\item{\code{stderr_log}}{standard error of \code{log(samples)}}
\item{\code{ir_95}}{the 95\% and 68\% interpercentile range}
\item{\code{ir_68}}{the 68\% interpercentile range}
\item{\code{samples}}{the Monte Carlo simulation}
}
Values will be in the unit specified by parameter \code{unit} or be equal to the unit of \code{x} if \code{x} is a \code{units} object.
}
\description{
Estimates from Monte Carlo Simulation
}
\details{
Equations of the form \eqn{X = A b^{n \pm \sigma}} create non-normal, left-skewed distributions.
Thus, it is recommended to report median and percentiles instead of mean, standard deviation and confidence intervals.
}
\examples{
mc_stats(rnorm(100, mean = 100), "MPa")
}
