% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strainrate.R
\name{strain_rate}
\alias{strain_rate}
\title{Strain rate}
\usage{
strain_rate(
  stress,
  temperature,
  fugacity = NULL,
  grainsize = NULL,
  pressure = NULL,
  sim = 1e+06,
  model = c("Hirth2001", "Paterson1990", "Kronenberg1984", "Luan1992", "Gleason2004",
    "Rutter2004", "Fukuda2018", "Richter2018", "Lu2019", "Tokle2019", "Lusk2021")
)
}
\arguments{
\item{stress}{Deviatoric stress in MPa or units object}

\item{temperature}{Temperature in Kelvin or units object}

\item{fugacity}{Water fugacity in MPa or units object}

\item{grainsize}{Grainsize in cm  or units object}

\item{pressure}{Pressure in MPa or units object}

\item{sim}{non-negative number. Number of Monte Carlo simulations}

\item{model}{character. Flow law.}
}
\value{
list. Strain rate in 1/s given as median of the Monte Carlo simulations \code{median}, mean (\code{mean}), standard error of \code{log(samples)} (\code{stderr_log}), the 95\% confidence interval of the mean (\code{conf.int}), and the Monte Carlo simulation (\code{samples})
}
\description{
Calculates strain rate from stress, temperature, and grain size.
Uses Monte Carlo sampling if flow model parameters contain uncertainties.
}
\details{
General flow law: \deqn{\dot{\epsilon} = A \sigma^n d^m f_{H_2O}^r \, e^{\left({\frac{-H}{RT}}\right)}}
}
\examples{
set.seed(20250411)
stress <- units::set_units(100, MPa)
temperature <- units::set_units(300, degC)
pressure <- units::set_units(400, MPa)
fugacity <- ps_fugacity(pressure, temperature)
strain_rate(stress = stress, temperature = temperature, model = "Paterson1990")
strain_rate(stress = stress, temperature = temperature, fugacity = fugacity, model = "Hirth2001")
strain_rate(stress = stress, temperature = temperature, fugacity = fugacity, model = "Rutter2004")
strain_rate(stress = stress, temperature = temperature, fugacity = fugacity, model = "Lu2019")
}
\references{
Hirth, G., Teyssier, C., & Dunlap, W. J. (2001). An evaluation of quartzite flow laws based on comparisons between experimentally and naturally deformed rocks. International Journal of Earth Sciences, 90(1), 77–87. https://doi.org/10.1007/s005310000152

Tokle, L., Hirth, G., & Behr, W. M. (2019). Flow laws and fabric transitions in wet quartzite. Earth and Planetary Science Letters, 505, 152–161. https://doi.org/10.1016/j.epsl.2018.10.017

Lusk, A. D. J., Platt, J. P., & Platt, J. A. (2021). Natural and Experimental Constraints on a Flow Law for Dislocation‐Dominated Creep in Wet Quartz. Journal of Geophysical Research: Solid Earth, 126(5), 1–25. https://doi.org/10.1029/2020JB021302
}
